Updated level scene, killplane
Add David's scene to the build settings so that it is selected from menu and goes to credits scene.

Added missing objects from game manager and end trigger so that the level can be completed in David's scene.

Updated killplane script so that it tracks the player's x position as they move around the level.